backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devscope-backend"
version = "0.1.0"
description = "DevScope — Repository Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "DevScope Team" }]
keywords = ["github", "metrics", "analytics", "bigquery"]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "psycopg[binary]==3.18.0",
    "alembic==1.13.0",
    "redis==5.0.1",
    "celery==5.3.4",
    "google-cloud-bigquery==3.14.0",
    "google-cloud-pubsub==2.18.4",
    "PyJWT==2.8.1",
    "cryptography==41.0.7",
    "httpx==0.25.2",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "ruff==0.1.8",
    "black==23.12.0",
    "isort==5.13.2",
    "mypy==1.7.1",
    "types-redis==4.6.0.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "RUF", "B", "ANN", "C4"]
ignore = ["E501", "ANN101"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
